requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
boto3==1.40.0
botocore==1.40.0
channels==4.3.0
channels_redis==4.3.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
jmespath==1.0.1
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
s3transfer==0.13.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
